SB-311 (refactor) : 중고 조회 응답 형식 변경 및 상세조회 api 변경
SB-311 (refactor) : 중고 조회 응답 형식 변경 및 상세조회 api 변경

1. UsedGoodsDetailResponse
- Date Pattern 변경
2. UsedGoodsSearchResponse
- Date Pattern 변경
3. 상세조회 api
- /api/usedgoods/{usedGoodsId} -> /open-api/usedgoods/{usedGoodsId}
