+++ warehouse/src/main/java/warehouse/domain/usedgoods/controller/UsedGoodsOpenApiController.java
@@ -9,10 +9,12 @@ import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import warehouse.domain.usedgoods.business.UsedGoodsBusiness;
 import warehouse.domain.usedgoods.controller.model.request.SearchCondition;
+import warehouse.domain.usedgoods.controller.model.response.UsedGoodsDetailResponse;
 import warehouse.domain.usedgoods.controller.model.response.UsedGoodsSearchResponse;
 
 @RestController
@@ -30,6 +32,13 @@ public class UsedGoodsOpenApiController {
     ) {
         List<UsedGoodsSearchResponse> response = usedGoodsBusiness.usedGoodsSearchBy(condition,
             page);
+        return Api.OK(response);
+    }
+
+    @GetMapping("/{usedGoodsId}") // usedGoodsId 로 중고 상세 조회
+    @Operation(summary = "[중고 아이디로 상세 조회]")
+    public Api<UsedGoodsDetailResponse> getUsedGoodsDetail(@PathVariable Long usedGoodsId) {
+        UsedGoodsDetailResponse response = usedGoodsBusiness.getUsedGoodsDetail(usedGoodsId);
         return Api.OK(response);
     }
