+++ warehouse/src/main/java/warehouse/domain/usedgoods/controller/model/response/UsedGoodsDetailResponse.java
@@ -1,5 +1,6 @@
 package warehouse.domain.usedgoods.controller.model.response;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -19,6 +20,7 @@ public class UsedGoodsDetailResponse {
 
     private String description;
 
+    @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
     private LocalDateTime postedAt;
 
     private GoodsResponse goods;
