+++ warehouse/src/main/java/warehouse/domain/usedgoods/controller/model/response/UsedGoodsSearchResponse.java
@@ -1,5 +1,6 @@
 package warehouse.domain.usedgoods.controller.model.response;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import db.domain.usedgoods.enums.UsedGoodsStatus;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
@@ -20,6 +21,7 @@ public class UsedGoodsSearchResponse {
 
     private int price;
 
+    @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
     private LocalDateTime postedAt;
 
     private UsedGoodsStatus status;
